LayoutTests/platform/ios/imported/w3c/web-platform-tests/css/css-borders/border-width-rounding.tentative-expected.txt

Test passes if border widths are rounded up when they are greater than 0 and less than 1, and rounded down when they are greater than 1.


FAIL Test that border widths are rounded up when they are greater than 0px but less than 1px, and rounded down when they are greater than 1px. assert_equals: expected "1px" but got "0.5px"
